Reorganize code for ease of maintenance
